feat: write ScrapeRules hashes to databases
